refactor: move routing config files into routingConfigs/ subdirectory
Mirrors the llmAcceleratorConfigs/ and topologyConfigs/ layout ahead of the tab
migration. Pure move plus relative-import depth fixes; no behavior change.

RHOAIENG-80020
